[ADD] Timeout socket error handled [FIX] Progress bar at the top of the fragment [ADD] Endless list: gets 20 rows each time

+++ app/src/main/java/es/agustruiz/deadpoolcomics/data/model/marvel/ComicResultMarvel.java
@@ -19,7 +19,7 @@ public class ComicResultMarvel {
     private String title;
     @SerializedName("issueNumber")
     @Expose
-    private Integer issueNumber;
+    private Float issueNumber;
     @SerializedName("variantDescription")
     @Expose
     private String variantDescription;
@@ -120,11 +120,11 @@ public class ComicResultMarvel {
         this.title = title;
     }
 
-    public Integer getIssueNumber() {
+    public Float getIssueNumber() {
         return issueNumber;
     }
 
-    public void setIssueNumber(Integer issueNumber) {
+    public void setIssueNumber(Float issueNumber) {
         this.issueNumber = issueNumber;
     }
